refactor: export signIn style in external file
